fix: resolve tree-sitter Node::child() usize/u32 mismatch workspace-wide
tree-sitter 0.26.11 changed Node::child() to take u32 instead of usize.
Every 0..node.child_count() loop feeding .child(i) across parser, indexer,
checker, cli, and lsp crates needed its index cast to u32 at the source
(for i in 0..count() as u32) rather than per-call-site casts.

Also fixes two clippy::too_many_lines violations that only surfaced once
the workspace could compile far enough for clippy to reach them:
extracts collect_enum_member_tokens() out of semantic_tokens::collect_tokens
and collect_checker_diags() out of cli::commands::check::run().

cargo fmt applied throughout.

+++ crates/lsp/src/document_links.rs
@@ -1,7 +1,7 @@
 use gdscript_parser::ParsedDocument;
 use tower_lsp::lsp_types::{DocumentLink, Position, Range, Url};
 
-/// Find all `res://` string literals in a GDScript document and return them
+/// Find all `res://` string literals in a `GDScript` document and return them
 /// as [`DocumentLink`]s that resolve to absolute filesystem paths.
 ///
 /// Only emits links for paths that exist on disk.
@@ -53,7 +53,7 @@ fn collect_links(
         return; // Don't recurse into string children.
     }
 
-    for i in 0..node.child_count() {
+    for i in 0..node.child_count() as u32 {
         let Some(child) = node.child(i) else { continue };
         collect_links(&child, source, workspace_root, out);
     }
@@ -61,9 +61,9 @@ fn collect_links(
 
 #[cfg(test)]
 mod tests {
-    use std::fs;
-    use gdscript_parser::parse::parse;
     use super::*;
+    use gdscript_parser::parse::parse;
+    use std::fs;
 
     #[test]
     fn res_path_that_exists_produces_link() {
